+++ skyportal/__init__.py
@@ -1,5 +1,26 @@
 __version__ = "1.4.0"
 
+# Monkey-patch matplotlib for simsurvey compatibility with matplotlib >= 3.8
+# simsurvey imports matplotlib.docstring which was renamed to matplotlib._docstring
+import matplotlib
+
+if not hasattr(matplotlib, "docstring"):
+    matplotlib.docstring = matplotlib._docstring
+
+# Register numpy types with psycopg2 for numpy 2.x compatibility
+# numpy 2 scalars no longer inherit from Python builtins, so psycopg2
+# needs explicit adapters to serialize them in SQL parameters
+import numpy as np
+from psycopg2.extensions import AsIs, register_adapter
+
+
+def _adapt_numpy_scalar(val):
+    return AsIs(repr(float(val)))
+
+
+for _np_type in [np.float64, np.float32, np.int64, np.int32, np.bool_]:
+    register_adapter(_np_type, _adapt_numpy_scalar)
+
 if "dev" in __version__:
     # Append last commit date and hash to dev version information, if available
